requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
pytest==8.3.3
httpx==0.27.2
python-dotenv==1.0.0
